bigvoto/pyproject.toml
[tool.poetry]
name = "bigvoto"
version = "0.1.0"
description = ""
authors = ["alexandrejastrow <alexandre.jastrow@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "^1.4.33"
alembic = "^1.7.7"
psycopg2-binary = "^2.9.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
asyncpg = "^0.25.0"
email-validator = "^1.1.3"
fastapi-mail = "^1.0.6"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
